# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blossom Buddy API",
        "description": "API pour la gestion des plantes",
        "contact": {
            "email": "admin@blossombuddy.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/plants/search": {
            "get": {
                "tags": [
                    "Plant Search"
                ],
                "summary": "Recherche de plantes par nom français ou anglais",
                "operationId": "6342a8e6974f612fd87cae941e7d4001",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Nom de la plante à rechercher",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "monstera"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre maximum de résultats",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 50,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultats de la recherche",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "plants": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "common_name": {
                                                        "type": "string",
                                                        "example": "Monstera deliciosa"
                                                    },
                                                    "french_name": {
                                                        "type": "string",
                                                        "example": "Monstera"
                                                    },
                                                    "alternative_names": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "watering_general_benchmark": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "query": {
                                            "type": "string",
                                            "example": "monstera"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Paramètre de recherche manquant"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/plants/autocomplete": {
            "get": {
                "tags": [
                    "Plant Search"
                ],
                "summary": "Autocomplétion pour les noms de plantes",
                "operationId": "4f49aca2060277b311944bf31122feca",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Début du nom de plante à compléter",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "mon"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre maximum de suggestions",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5,
                            "maximum": 20,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suggestions d'autocomplétion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "suggestions": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID de la plante (si trouvée en base)",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "text": {
                                                        "type": "string",
                                                        "example": "Monstera"
                                                    },
                                                    "secondary": {
                                                        "type": "string",
                                                        "example": "Monstera deliciosa"
                                                    },
                                                    "source": {
                                                        "type": "string",
                                                        "enum": [
                                                            "database",
                                                            "dictionary"
                                                        ]
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "query": {
                                            "type": "string",
                                            "example": "mon"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Paramètre de recherche invalide"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/plants/find-or-suggest": {
            "post": {
                "tags": [
                    "Plant Search"
                ],
                "summary": "Trouve une plante ou suggère des alternatives",
                "operationId": "77d426357e7d4ff43cbc022b2824c21f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "french_name": {
                                        "description": "Nom français de la plante",
                                        "type": "string",
                                        "example": "monstera"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plante trouvée ou suggestions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "found": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "plant": {
                                                    "type": "object"
                                                },
                                                "confidence": {
                                                    "type": "string",
                                                    "example": "high"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "found": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "suggestions": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Nom de plante manquant"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/plants/{id}/french-names": {
            "put": {
                "tags": [
                    "Plant Search"
                ],
                "summary": "Met à jour les noms français d'une plante",
                "operationId": "93156440d6d25ab23024e1db2d54c79c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la plante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "french_name": {
                                        "description": "Nom français principal",
                                        "type": "string",
                                        "example": "Monstera"
                                    },
                                    "alternative_names": {
                                        "description": "Noms alternatifs et synonymes",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Faux philodendron",
                                            "Plante gruyère"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Noms français mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Noms français mis à jour avec succès"
                                        },
                                        "plant": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plante non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}